--- android/bip32ed25519/build.gradle
@@ -1,67 +1,0 @@
-version = "0.1.0"
-group = "com.algorandfoundation.bip32ed25519"
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-
-android {
-    compileSdkVersion 34
-    buildToolsVersion "33.0.3"
-
-    buildFeatures {
-        viewBinding = true
-    }
-
-    defaultConfig {
-        namespace "com.algorandfoundation.bip32ed25519"
-        minSdkVersion 19
-        targetSdkVersion 33
-        versionCode 16
-        versionName "0.1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        multiDexEnabled true
-    }
-
-    lint {
-        baseline = file("lint-baseline.xml")
-    }
-
-    sourceSets {
-        main {
-            jni.srcDirs = []
-            jniLibs.srcDirs = ['src/main/jniLibs']
-        }
-    }
-
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_17
-        targetCompatibility JavaVersion.VERSION_17
-    }
-}
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'net.java.dev.jna:jna:5.12.1'
-    implementation 'com.goterl:resource-loader:2.0.2'
-    implementation 'cash.z.ecc.android:kotlin-bip39:1.0.7'
-    implementation 'net.pwall.json:json-kotlin-schema:0.46'
-    implementation 'com.algorand:algosdk:2.4.0'
-    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3'
-    implementation 'org.jetbrains.kotlinx:kotlinx-serialization-core:1.6.3'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.16.1'
-    implementation 'com.fasterxml.jackson.core:jackson-core:2.16.1'
-    implementation 'org.msgpack:jackson-dataformat-msgpack:0.9.8'
-    implementation 'com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.16.1'
-    testImplementation 'org.jetbrains.kotlin:kotlin-test-junit5'
-    testImplementation libs.junit.jupiter.engine
-    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    api libs.commons.math3
-    implementation libs.guava
-}
-
-java {
-    toolchain {
-        languageVersion.set(JavaLanguageVersion.of(17))
-    }
-}
