--- android/gradle/libs.versions.toml
@@ -1,23 +1,0 @@
-# This file was generated by the Gradle 'init' task.
-# https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
-
-[versions]
-commons-math3 = "3.6.1"
-guava = "32.1.2-jre"
-junit-jupiter-engine = "5.10.0"
-junit = "4.13.2"
-runner = "1.0.2"
-espresso-core = "3.0.2"
-appcompat = "1.6.1"
-
-[libraries]
-commons-math3 = { module = "org.apache.commons:commons-math3", version.ref = "commons-math3" }
-guava = { module = "com.google.guava:guava", version.ref = "guava" }
-junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junit-jupiter-engine" }
-junit = { group = "junit", name = "junit", version.ref = "junit" }
-runner = { group = "com.android.support.test", name = "runner", version.ref = "runner" }
-espresso-core = { group = "com.android.support.test.espresso", name = "espresso-core", version.ref = "espresso-core" }
-androidx_appcompat = { module = "androidx.appcompat:appcompat", version.ref = "appcompat" }
-
-[plugins]
-jvm = { id = "org.jetbrains.kotlin.jvm", version = "1.9.20" }
