+++ androidModule/src/main/jniLibs/arm64-v8a/libsodium.so
