+++ androidModule/src/main/jniLibs/armeabi-v7a/libsodium.so
