+++ androidModule/src/main/jniLibs/armeabi/libsodium.so
