+++ androidModule/src/main/jniLibs/x86/libsodium.so
