+++ androidModule/src/main/jniLibs/x86_64/libsodium.so
