+++ androidModule/src/main/kotlin/bip32ed25519/Bip32Ed25519Android.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) Algorand Foundation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package bip32ed25519
+
+import com.goterl.lazysodium.LazySodiumAndroid
+import com.goterl.lazysodium.SodiumAndroid
+
+class Bip32Ed25519Android(private var seed: ByteArray) : Bip32Ed25519Base(seed) {
+    override val lazySodium = LazySodiumAndroid(SodiumAndroid())
+}
