+++ build.gradle.kts
@@ -1,0 +1,15 @@
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+plugins {
+    id("com.android.application") version "8.1.4" apply false
+    id("org.jetbrains.kotlin.android") version "1.9.0" apply false
+    id("com.android.library") version "8.1.4" apply false
+    `java-library`
+    kotlin("jvm") version "1.9.22"
+}
+
+
+allprojects {
+    tasks.withType<JavaCompile> {
+        options.encoding = "UTF-8"
+    }
+}
