--- desktop/bip32ed25519/build.gradle.kts
@@ -1,86 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Kotlin library project to get you started.
- * For more details on building Java & JVM projects, please refer to https://docs.gradle.org/8.5/userguide/building_java_projects.html in the Gradle documentation.
- */
-
-version = "0.1.0"
-group = "com.algorandfoundation.bip32ed25519"
-
-plugins {
-    kotlin("jvm") version "1.9.22"
-    kotlin("plugin.serialization") version "1.9.22"
-    java
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-
-}
-
-repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
-}
-
-
-dependencies {
-    // Lazy Sodium Fork + dependencies:
-    implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar"))))
-    implementation("net.java.dev.jna:jna:5.12.1")
-    implementation("com.goterl:resource-loader:2.0.2")
-
-    // Bip39 implementation
-    implementation("cash.z.ecc.android:kotlin-bip39:1.0.7") 
-
-    // For data validation
-    implementation("net.pwall.json:json-kotlin-schema:0.46")
-    implementation("com.algorand:algosdk:2.4.0")
-
-    // https://mvnrepository.com/artifact/org.jetbrains.kotlinx
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.6.3")
-    implementation("org.jetbrains.kotlinx:kotlinx-serialization-core:1.6.3")
-    // https://mvnrepository.com/artifact/com.fasterxml.jackson.core
-    implementation("com.fasterxml.jackson.core:jackson-databind:2.16.1")
-    implementation("com.fasterxml.jackson.core:jackson-core:2.16.1")
-    implementation("org.msgpack:jackson-dataformat-msgpack:0.9.8")
-    implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.16.1")
-
-    
-
-
-    // Use the Kotlin JUnit 5 integration.
-    testImplementation("org.jetbrains.kotlin:kotlin-test-junit5")
-
-    // Use the JUnit 5 integration.
-    testImplementation(libs.junit.jupiter.engine)
-
-    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
-
-    // This dependency is exported to consumers, that is to say found on their compile classpath.
-    api(libs.commons.math3)
-
-    // This dependency is used internally, and not exposed to consumers on their own compile
-    // classpath.
-    implementation(libs.guava)
-}
-
-// Apply a specific Java toolchain to ease working on different environments.
-java { toolchain { languageVersion.set(JavaLanguageVersion.of(17)) } }
-
-// Run ./gradlew test to execute tests not requiring an Algorand Sandbox network
-tasks.named<Test>("test") {
-    // Use JUnit Platform for unit tests.
-    useJUnitPlatform{
-        excludeTags("sandbox")
-    }
-    testLogging.showStandardStreams = true
-}
-
-// Run ./gradlew testWithAlgorandSandbox to run tests that interact with an Algorand Sandbox network (e.g. Algokit Localnet)
-tasks.register<Test>("testWithAlgorandSandbox") {
-    useJUnitPlatform()
-    testLogging.showStandardStreams = true
-}
