--- desktop/gradle/libs.versions.toml
@@ -1,15 +1,0 @@
-# This file was generated by the Gradle 'init' task.
-# https://docs.gradle.org/current/userguide/platforms.html#sub::toml-dependencies-format
-
-[versions]
-commons-math3 = "3.6.1"
-guava = "32.1.2-jre"
-junit-jupiter-engine = "5.10.0"
-
-[libraries]
-commons-math3 = { module = "org.apache.commons:commons-math3", version.ref = "commons-math3" }
-guava = { module = "com.google.guava:guava", version.ref = "guava" }
-junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junit-jupiter-engine" }
-
-[plugins]
-jvm = { id = "org.jetbrains.kotlin.jvm", version = "1.9.20" }
