+++ jvmModule/src/test/resources/auth.request.json
