+++ jvmModule/src/test/resources/msg.schema.json
