+++ sharedModule/src/main/kotlin/bip32ed25519/utils.kt
